Add e1000 driver for host devices for bonding
Change the default virtio driver to e1000 for all interfaces on boxes not containing the name "vx", ie host machines.  This is needed to allow creation of compatible bonds on hosts with Cumulus VX devices using the default virtio driver.  The issue was tested with the virtio driver loaded on a Ubuntu machine where it is not setting correct L1 parameters for a bond interface.  NOTE: This fix will change the driver to e1000 for all interfaces on the host.  Another way to fix is to change only the network facing ports but waiting on fix for: https://github.com/vagrant-libvirt/vagrant-libvirt/issues/634