Delete "phantom nics" 9 and 10 in CumulusVX for Internet device
The CumulusVX boxes seem to have two nics that "show up" in positions 9 and 10.
If these nics are not connected to any nets, they try to connect to the
VirtualBox "intnet", which can cause conflicts if multiple users are on the same
system.

This fix also removes the "virtio" designation for the Internet device nics
since they are assigned by default to all CumulusVX nics.