Clean up whitespace
There were places in VM definitions where either the command or delimiter
indentation was not correct, thus making it awkward to understand scoping.  In
addition, the use of "extra lines" was inconsistent and
non-intuitive... multiple blank lines occured inside of VM definitions, but only
a single blank line was used between VM definitions.  Cleaned all of this up to
make the file more readable.