Use $USER as the default workbench-id (wbid) for multi-user systems
There is a designator "wbid" that allows for poor-man's name-space separation
when multiple users run this Vagrantfile on the same system.  Unfortunately,
this requires that users think to edit the Vagrantfile and know what values
others might use.

This changes the default wbid from "1" to the $USER environmental variable so
that, in many cases, the users will be able to get by without modifying
anything.