Make sure that the "backup" Vagrantfile-vbox is the same as Vagrantfile
The repository defaults to using VirtualBox as the Vagrant provider; however, it
provides Vagrantfiles that allow users to pick different providers.  As part of
this helper function, the repository makes it easy to get back to the original
by providing a copy of the VirtualBox Vagrantfile.

Based on both slippage over time as well as recent changes to the Vagrantfile,
just need to keep these in sync.